Script to generate a trace from artifact files
ovn-trace
OUT=output-dir ovn-trace

Trace is done in current directory which must contain previously
downloaded CI artifacts from an openshift-ovn-kubernetes cluster.
The output trace is in ./trace_out or content of OUT.

The *-journal and openshift-ovn-namespace log files are processed.
The hyperkube SyncLoop and errors and crio are extracted from the
journals.
The whole current and previous logs are processed.

Each log line gets a common prefix that consists of
- timestamp in a common format
- node name
- pod name
- container name
The remainder of the line is not changed.

All of the lines are sorted into a time sequence trace.
The trace is analyzed for ovn container bringup and reported in another
file.

This is useful for analyzing ci test failures until the must-gather
version is available. Its also a basis for additional experimenting.

Signed-off-by: Phil Cameron <pcameron@redhat.com>